feat: backend data structures change

+++ backend/PhotoLibraryBackend.Common/Model/ImporterReport.cs
@@ -1,5 +1,5 @@
 ﻿namespace PhotoLibraryBackend.Common;
 
-public record ImporterReport(int Timestamp, ImporterReportSeverity Severity, string Message) {
+public record ImporterReport(long Timestamp, ImporterReportSeverity Severity, string Message) {
     public long Id { get; set; }
 }
